refactor: move `PopupPage` component to a new file
