+++ src/assets/ts/components/popup-page.tsx
@@ -1,7 +1,6 @@
 /* global EventListener */
-import ReactDOM from 'react-dom/client';
 import { useEffect, useState } from 'react';
-import { Provider, useDispatch, useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import 'bootstrap-icons/font/bootstrap-icons.scss';
 
 import '@/scss/style.scss';
@@ -13,12 +12,11 @@ import { Proposal } from '@/components/proposal';
 import { SignInSuccess } from '@/components/sign-in-success';
 import { Settings } from '@/components/settings';
 
-import { store } from '@/store';
 import { signIn } from '@/features/popup-slice';
 import { RootState, AppDispatch } from '@/store';
 import { CustomEvent } from '@/types/core';
 
-const PopupPage: React.FC = () => {
+export const PopupPage: React.FC = () => {
   const dispatch = useDispatch<AppDispatch>();
 
   const {
@@ -65,13 +63,3 @@ const PopupPage: React.FC = () => {
     </>
   );
 };
-
-const container = document.getElementById('app');
-if (!container) throw new Error('Failed to find the app element');
-
-const root = ReactDOM.createRoot(container);
-root.render(
-  <Provider store={store}>
-    <PopupPage />
-  </Provider>,
-);
